Removes all use of unsigned long and some use of long.

+++ common/bfiofile.h
@@ -49,7 +49,7 @@ public:
     BufferIOFileClass(void);
     virtual ~BufferIOFileClass(void);
 
-    bool Cache(long size = 0, void* ptr = NULL);
+    bool Cache(int size = 0, void* ptr = NULL);
     void Free(void);
     bool Commit(void);
     virtual char const* Set_Name(char const* filename);
@@ -57,10 +57,10 @@ public:
     virtual int Is_Open(void) const;
     virtual int Open(char const* filename, int rights = READ);
     virtual int Open(int rights = READ);
-    virtual long Read(void* buffer, long size);
-    virtual long Seek(long pos, int dir = SEEK_CUR);
-    virtual long Size(void);
-    virtual long Write(void const* buffer, long size);
+    virtual int Read(void* buffer, int size);
+    virtual int Seek(int pos, int dir = SEEK_CUR);
+    virtual int Size(void);
+    virtual int Write(void const* buffer, int size);
     virtual void Close(void);
 
     enum
@@ -80,14 +80,14 @@ private:
 
     void* Buffer;
 
-    long BufferSize;
-    long BufferPos;
-    long BufferFilePos;
-    long BufferChangeBeg;
-    long BufferChangeEnd;
-    long FileSize;
-    long FilePos;
-    long TrueFileStart;
+    int BufferSize;
+    int BufferPos;
+    int BufferFilePos;
+    int BufferChangeBeg;
+    int BufferChangeEnd;
+    int FileSize;
+    int FilePos;
+    int TrueFileStart;
 };
 
 #endif
