+++ common/combuf.h
@@ -58,9 +58,9 @@ typedef struct
 {
     unsigned int IsActive : 1; // 1 = this entry is ready to be processed
     unsigned int IsACK : 1;    // 1 = ACK received for this packet
-    unsigned long FirstTime;   // time this packet was first sent
-    unsigned long LastTime;    // time this packet was last sent
-    unsigned long SendCount;   // # of times this packet has been sent
+    unsigned int FirstTime;    // time this packet was first sent
+    unsigned int LastTime;     // time this packet was last sent
+    unsigned int SendCount;    // # of times this packet has been sent
     int BufLen;                // size of the packet stored in this entry
     char* Buffer;              // the data packet
     int ExtraLen;              // size of extra data
@@ -112,7 +112,7 @@ public:
         return (MaxSend);
     }                                   // max # send queue entries
     SendQueueType* Get_Send(int index); // random access to queue
-    unsigned long Send_Total(void)
+    unsigned int Send_Total(void)
     {
         return (SendTotal);
     }
@@ -131,7 +131,7 @@ public:
         return (MaxReceive);
     }                                         // max # recv queue entries
     ReceiveQueueType* Get_Receive(int index); // random access to queue
-    unsigned long Receive_Total(void)
+    unsigned int Receive_Total(void)
     {
         return (ReceiveTotal);
     }
@@ -139,10 +139,10 @@ public:
     /*
     ....................... Response time routines ........................
     */
-    void Add_Delay(unsigned long delay);   // accumulates response time
-    unsigned long Avg_Response_Time(void); // gets mean response time
-    unsigned long Max_Response_Time(void); // gets max response time
-    void Reset_Response_Time(void);        // resets computations
+    void Add_Delay(unsigned int delay);   // accumulates response time
+    unsigned int Avg_Response_Time(void); // gets mean response time
+    unsigned int Max_Response_Time(void); // gets max response time
+    void Reset_Response_Time(void);       // resets computations
 
     /*
     ........................ Debug output routines ........................
@@ -166,17 +166,17 @@ private:
     /*
     ....................... Response time variables .......................
     */
-    unsigned long DelaySum;  // sum of last 4 delay times
-    unsigned long NumDelay;  // current # delay times summed
-    unsigned long MeanDelay; // current average delay time
-    unsigned long MaxDelay;  // max delay ever for this queue
+    unsigned int DelaySum;  // sum of last 4 delay times
+    unsigned int NumDelay;  // current # delay times summed
+    unsigned int MeanDelay; // current average delay time
+    unsigned int MaxDelay;  // max delay ever for this queue
 
     /*
     ........................ Send Queue variables .........................
     */
     SendQueueType* SendQueue; // incoming packets
     int SendCount;            // # packets in the queue
-    unsigned long SendTotal;  // total # added to send queue
+    unsigned int SendTotal;   // total # added to send queue
     int* SendIndex;           // array of Send entry indices
 
     /*
@@ -184,7 +184,7 @@ private:
     */
     ReceiveQueueType* ReceiveQueue; // outgoing packets
     int ReceiveCount;               // # packets in the queue
-    unsigned long ReceiveTotal;     // total # added to receive queue
+    unsigned int ReceiveTotal;      // total # added to receive queue
     int* ReceiveIndex;              // array of Receive entry indices
 
     /*
